// src/api/schemas/getClientesSort.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */

export type GetClientesSort = typeof GetClientesSort[keyof typeof GetClientesSort];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetClientesSort = {
  nombre: 'nombre',
  apellido: 'apellido',
  nacimiento: 'nacimiento',
  sexo: 'sexo',
  pais: 'pais',
  pasaporte: 'pasaporte',
  email: 'email',
} as const;