// src/api/schemas/getClientesSortOrder.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */

export type GetClientesSortOrder = typeof GetClientesSortOrder[keyof typeof GetClientesSortOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetClientesSortOrder = {
  ASC: 'ASC',
  DESC: 'DESC',
} as const;